+++ templates/components/layout.html
@@ -32,6 +32,7 @@
         {% include 'components/navigation.html' %}
     </header>
     {% block contents %}
+    
     {% endblock %}
     <footer>
         {% include 'components/navigation.html' %}
